backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
python-dotenv
supabase
Pillow
requests
gradio_client
langgraph
langchain
langchain-google-genai
